Use Ncurses.cbreak to support single char input
Without this modification when ffi-ncurses is used in JRuby
only the first of the following two agree invocations
returns after a single char input. The second invocation
waits for a newline.

  agree("Are you sure you want to do this?  (y/n) ", true)
  agree("Are you really sure you want to do this?  (y/n) ", true)

This patch fixes the problem.

I tried to write the following as a test:

def test_agree_singlechar_twice
  @input << "yy"
  @input.rewind
  assert_equal(true, @terminal.agree("Yes or no?  "), true)
  assert_equal(true, @terminal.agree("Yes or no?  "), true)
end

But an: EOFError: The input stream is exhausted.
error was generated in MRI and JRuby.