Handle empty items consistently in HighLine#list
Without this patch HighLine#list throws an error when called with an
empty items array and a :columns_across or :columns_down mode. For
example HighLine.new.list([], :columns_across) results in:

  NoMethodError: private method `gsub' called for nil:NilClass
    actual_length at .../highline-1.6.2/lib/highline.rb:761
             list at .../highline-1.6.2/lib/highline.rb:394

This patch fixes this error so that HighLine#list always returns an
empty string when called with an empty items array, regardless of mode.