Modified HighLine#say to better handle colorized statements.
The original HighLine#say method will correctly not emit a newline with the statement (displaying with print and flush) when it ends in whitespace. However, if one creates a string ending in whitespace and then uses HighLine#color to colorize it, HighLine#say will emit the newline, which is rarely-if-ever the intended behavior. I have modified the method to cause it to print and flush (no newline) when the statement string ends with whitespace directly preceding a color code, by way of a regex.