Use the configured streams on JRuby with JLine.
Before this commit, the input and output parameters to HighLine.new
were ignored for #ask (both for specifying the input stream and for
displaying the prompt).

This commit fixes issue #31, which was introduced by PR #27. Since
then, the HighLine tests have been unable to run on JRuby. With this
change, the tests run on JRuby 1.7.3 but many of them fail. No tests
fail on JRuby 1.7.3 using the version of HighLine from immediately
before #27 was merged, so it seems likely the failures were caused
by #27 also.

This commit stops using `SystemExtensions#initialize` to set up JLine.
It would have been possible to pass the input and output parameters to
the `super` call, but that seemed to me like an abuse of inheritance.
Instead, following the pattern of the rest of SystemExtensions, I
added a method called `initialize_system_extensions` which is defined
on a per-platform basis (currently for JRuby only) and invoked by
`HighLine#initialize` when present.