Make methods in SystemExtensions public
Most of the methods rely on instance variables, so it does not
make sense that they should be available as module methods.
But when they are included into the HighLine class, they
are private, so no useful outside of the class.

Making the methods public doesn't seem to have a negative
effect on test results, and it allows the methods to be called
on instances of the HighLine class.