Improve on #110 - STDIO.external_encoding returning nil
The last PR #110 passes all the tests because
StringIO#external_encoding answers correctly with a valid encoding.

Although at the field, @output will be pointing to STDIO.
STDIO.external_encoding returns nil as stated in
http://www.ruby-doc.org/core-2.1.0/IO.html#method-i-external_encoding
  > Returns the Encoding object that represents the encoding of the file.
  > If io is write mode and no encoding is specified, returns nil.

Encoding.default_external is probably the correct way to
get the job done.

Improve on #110
Related to #108