Completely remove @question inst var from HighLine objects
... and fix tests accordling.

Removing unecessary _state_ handling (inst vars) is a step
toward a thiner HighLine class.

Question shouldn't be a HighLine's _state_ because the **same**
HighLine instance could be able to ask several questions.

(At the previous code @question was a kind of _current_question_ state.
But this is not really necessary currently.)