Add some semi-automatic 'acceptance' tests to HighLine (yup)
These 'acceptance' tests guide the user through some steps
asking if they SEE what they should be seeing.

It summarizes the answers and adds some debug environment
information that can be sent to HighLine contributors.

It'll be a complementary way of testing HighLine without the caveats
of using a fake input/output object like StringIO as we do
on automatic unit tests.