Fix agree validation to only accept "yes" or "no"
Although the documentation for agree states that it only accepts
yes/no/y/n case-insensitively, the Regexp allowed any string which
started with /y/i or ended with /no?/i and interpreted any starting with
'y' as true.  Fix the Regexp to properly exclude the start/end markers
from the alternation and add a test value to confirm it works.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>