Raise Ruby version requirement to 3.0
Ruby 2.7 is tagged End of Life (not even security maintanence) since this month.
We (Highline) traditionally try to maintain compatibility with old rubies.
But rubies bellow 2.7.1 lack "io/console" or they have incompatibilities.
(as discussed in PR #259)

So, this EOL of 2.7 signal to us that is time for us to move on.
People relying on old versions of Highline will probably face no problem with it.
But we will be free to start removing old code that is necessary only for
retrocompatibility in newer versions.